requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-messaging-webpubsubservice
azure-data-tables
requests_async
azure-ai-translation-document
azure-storage-blob
pydub
azure-identity
azure-mgmt-media
python-dotenv
aiohttp
azure-common
msrestazure
